pyproject.toml
[tool.poetry]
name = "runekit"
version = "0.2.0"
description = ""
authors = ["Chris Moran <chris@chrismoran.com>"]
license = "GPL"
include = ["runekit/_resources.py"]

[tool.poetry.dependencies]
python = "^3.9,<3.11"
requests = "^2.26.0"
PySide2 = "^5.15.2"
Pillow = "^8.3.1"
sysv-ipc = [{ version="^1.1.0", platform="linux" }]
xcffib = [{ version="^0.11.1", platform="linux" }]
pyobjc-framework-Quartz = [{ version="7.3", platform="darwin" }]
pyobjc-framework-ApplicationServices = [{ version="7.3", platform="darwin" }]
pyobjc-framework-CoreText = [{ version="7.3", platform="darwin" }]
click = "^8.0.1"
psutil = "^5.8.0"
opencv-python-headless = "^4.5.3.56"
pytesseract = "^0.3.9"
scikit-image = "^0.19.3"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pyinstaller = [{ version="^4.5.1", platform="darwin" }]

[tool.poetry.scripts]
runekit = 'runekit.main:main'
